scripts/requirements-cluster.txt
# Requirements for 2D MetaWorld training on cluster
# Minimal set - excludes: pytorch3d, sapien, mplib, open3d
# PyTorch is pre-installed in the base container image
# Third-party needed: gym-0.21.0, Metaworld, r3m, mujoco-py-2.1.2.14

# =====================================================
# Core dependencies (relaxed numpy for numba compat)
# =====================================================
setuptools>=70.0.0
Cython==0.29.35
numpy==1.26.4
scipy>=1.10.0
scikit-learn>=1.3.0

# =====================================================
# Visualization
# =====================================================
matplotlib
seaborn
plotly
kaleido
imageio
moviepy
opencv-python==4.5.5.64

# =====================================================
# Data handling
# =====================================================
h5py
zarr==2.12.0
dill==0.3.5.1

# =====================================================
# Configuration and logging
# =====================================================
omegaconf
hydra-core==1.2.0
hydra-colorlog
wandb
tqdm
tabulate
termcolor

# =====================================================
# RL/Robotics (for dm_control and gymnasium)
# =====================================================
dm_control
gymnasium==0.29.1
pyglet<2

# =====================================================
# Deep learning utilities
# =====================================================
torchdiffeq
torchsde
kornia
einops
timm
accelerate>=0.34.0
diffusers>=0.27.0

# =====================================================
# Transformers (for model encoders)
# =====================================================
transformers>=4.40.0
huggingface_hub>=0.25.0
safetensors
ftfy
regex
openai==1.54.3
azure==4.0.0
openai==1.54.3	
azure==4.0.0	
sentence-transformers
sentencepiece==0.2.0	
deepspeed==0.15.1	
# =====================================================
# Utilities
# =====================================================
pydantic
av
decorator
gpustat
natsort
transforms3d
trimesh
shortuuid
numba==0.61.2
gdown
tenacity
patchelf

# =====================================================
# Development
# =====================================================
ipdb
pytest

# =====================================================
# MuJoCo-py dependencies
# =====================================================
glfw>=1.4.0
cffi>=1.10
fasteners>=0.15
lockfile

# =====================================================
# Gym 0.21.0 dependencies
# =====================================================
cloudpickle>=1.2.0
lz4>=3.1.0
